+++ src/main/java/roomescape/controller/CreateReservationRequest.java
@@ -1,6 +1,8 @@
-package roomescape;
+package roomescape.controller;
 
 import java.time.LocalDateTime;
+import roomescape.domain.Reservation;
+import roomescape.util.CustomDateTimeFormatter;
 
 public record CreateReservationRequest(String date, String name, String time) {
