+++ src/main/java/roomescape/controller/CreateReservationResponse.java
@@ -1,4 +1,7 @@
-package roomescape;
+package roomescape.controller;
+
+import roomescape.domain.Reservation;
+import roomescape.util.CustomDateTimeFormatter;
 
 public record CreateReservationResponse(long id, String name, String date, String time) {
