+++ src/test/java/roomescape/CustomDateTimeFormatterTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.util.CustomDateTimeFormatter;
 
 class CustomDateTimeFormatterTest {
